AIEngineerAssignment-Virallens/q1_rag_orchestration/requirements.txt
# Core
langchain>=0.3.10
langchain-community>=0.3.10
langgraph>=0.2.48

# Vector DB
qdrant-client>=1.12.0
langchain-qdrant>=0.2.0

# LLM Providers
groq
google-generativeai
openrouter
ollama

# Embeddings & Reranking
sentence-transformers

# Cache
redis

# PDF Processing
pypdf
PyPDF2

# Utilities
python-dotenv
pydantic
pytest
pytest-asyncio
pytest-cov
pytest-mock
